Big update!
Called it good enough on the chaos mage tab (only needs UI help, turns out it's slower to input things than to pencil and paper. Was worthwhile to code to learn!)

Big add to the initiative tab. It works! It has animals! It has an update button (which should probably be advance...)